Fixed win32 squeshfuse_extract implementation to correctly set the directory bit when creating  symbolic link to a directory.
reverted changes to external dependencies project that attempted to prevent
unwanted rebuilds. For now, will just build all of the external dependencies, then
unload the project.

Added icons & version resources to the executables.

Added a custom icon to the wix installer.

+++ extract.c
@@ -170,24 +170,30 @@ int main(int argc, char *argv[]) {
                     fclose(f);
                     chmod (prefixed_path_to_extract, st.st_mode);
                 } else if (inode.base.inode_type == SQUASHFS_SYMLINK_TYPE){
-                    size_t size = strlen(trv.path)+1;
-                    /* char buf[size]; */
-					char *buf = malloc(size);
-					if (buf)
-					{
-						int ret = sqfs_readlink(&fs, &inode, buf, &size);
-						if (ret != 0)
-							die("sqfs_readlink error");
-						fprintf(stderr, "Symlink: %s to %s \n", prefixed_path_to_extract, buf);
-						unlink(prefixed_path_to_extract);
-						ret = symlink(buf, prefixed_path_to_extract);
-						if (ret != 0)
-							die("symlink error");
-						free(buf);
+                    size_t size = 0;
+					/* get the buffer size we'll need to read the link. */
+					int ret = sqfs_readlink(&fs, &inode, NULL, &size);
+					if (ret == SQFS_OK) {
+						char *buf = malloc(size);
+						if (buf)
+						{
+							int ret = sqfs_readlink(&fs, &inode, buf, &size);
+							if (ret != 0)
+								die("sqfs_readlink error");
+							fprintf(stderr, "Symlink: %s to %s \n", prefixed_path_to_extract, buf);
+							unlink(prefixed_path_to_extract);
+							ret = sqfs_symlink(&fs, buf, prefixed_path_to_extract);
+							if (ret != 0)
+								die("symlink error");
+							free(buf);
+						}
+						else
+						{
+							die("memory allocation failure");
+						}
 					}
-					else
-					{
-						die("memory allocation failure");
+					else {
+						die("sqfs_readlink error");
 					}
                 } else {
                     fprintf(stderr, "TODO: Implement inode.base.inode_type %i\n", inode.base.inode_type);
